# apps/api/src/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kubernal IDP API",
    "version": "0.1.0",
    "description": "Internal Developer Platform - Self-service application orchestrator",
    "contact": {
      "name": "Platform Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/v1/deployments/{id}/violations": {
      "post": {
        "tags": ["Deployments"],
        "summary": "Record policy violations for a deployment",
        "operationId": "recordViolations",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["violations"],
                "properties": {
                  "violations": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "policyId": { "type": "string", "format": "uuid" },
                        "policyName": { "type": "string" },
                        "severity": {
                          "type": "string",
                          "enum": ["critical", "high", "medium", "low"]
                        },
                        "message": { "type": "string" },
                        "resource": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Violations recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Deployment" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
...
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": { "type": "object" }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["success", "error"]
      },
...
      "Deployment": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "applicationId": { "type": "string", "format": "uuid" },
          "environmentId": { "type": "string", "format": "uuid" },
          "version": { "type": "string" },
          "commitSha": { "type": "string" },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "building",
              "deploying",
              "healthy",
              "failed",
              "rolled_back",
              "cancelled"
            ]
          },
          "trigger": { "type": "string", "enum": ["manual", "git_push", "scheduled", "rollback"] },
          "approvedById": { "type": "string", "format": "uuid", "nullable": true },
          "policyViolations": { "type": "array", "items": { "type": "object" } },
          "startedAt": { "type": "string", "format": "date-time" },
          "completedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "id",
          "applicationId",
          "environmentId",
          "version",
          "commitSha",
          "status",
          "trigger",
          "startedAt",
          "createdAt"
        ]
      },
...
    }
  },
...
}